requirements.txt
torch
torchvision
pandas
numpy
scikit-learn
xgboost
matplotlib
seaborn
Pillow
tqdm
openpyxl
requests
ydata-profiling
grad-cam